made the addToCart logic reuseAble
